feat(core/rust): decrease size of Qstr type to u16
which is perfectly legal now that we have Attribute type to cover the
ffi::qstr case.

+++ core/embed/rust/qstr.h
@@ -1,6 +1,6 @@
-#include <stddef.h>
+#include <stdint.h>
 
-enum Qstr : size_t {
+enum Qstr : uint16_t {
 
 // Copied from `vendor/micropython/py/qstr.h`:
